pic-chatter/backend/requirements.txt
flask==3.0.2 
openai==1.1.0
requests==2.28.1
fastapi==0.115.4
uvicorn==0.29.0
python-dotenv==1.0.1